+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-05-06 13:24:08
-// Commit: docs(perf): update readme and API contracts for perf
+// Updated at: 2026-05-06 18:17:29
+// Commit: feat(perf): integrate custom telemetry logger in perf
 
 class perfComponentState {
-  final String commitRef = '935';
-  final String lastMessage = 'docs(perf): update readme and API contracts for perf';
-  final DateTime timestamp = DateTime.parse('2026-05-06T13:24:08+05:00');
+  final String commitRef = '951';
+  final String lastMessage = 'feat(perf): integrate custom telemetry logger in perf';
+  final DateTime timestamp = DateTime.parse('2026-05-06T18:17:29+05:00');
 }